votes
git-svn-id: https://vote4music.googlecode.com/svn/trunk@56 f4e7aff0-2383-11df-8b8d-d39edfa62fdf

+++ play/app/controllers/Application.java
@@ -199,7 +199,7 @@ public class Application extends Controller {
 	public static void rate(String id, String rating) {
 		String albumId = id.substring(6);
 		Album album = Album.findById(Long.parseLong(albumId));
-		album.rate(Integer.parseInt(rating));
+		album.rate(Float.parseFloat(rating));
 		renderText(rating);
 	}
